+++ myserverweb/include/gzip.h
@@ -43,16 +43,18 @@ public:
 		u_long data_size;
 		u_long initialized;
 	};
+
 #ifdef GZIP_CHECK_BOUNDS	
-	static int compressBound(int size);
+	static u_long compressBound(int size);
 #endif
-	int updateCRC(char* buffer,int size);
-	int getFOOTER(char *str,int size);
-	int initialize();
-	int compress(char* in,u_long sizeIN,char *out,u_long sizeOUT);
-	int free();
-	int flush(char *out,u_long sizeOUT);
-	int getHEADER(char *buffer,u_long buffersize);
+
+	u_long updateCRC(char* buffer,int size);
+	u_long getFOOTER(char *str,int size);
+	u_long initialize();
+	u_long compress(char* in,u_long sizeIN,char *out,u_long sizeOUT);
+	u_long free();
+	u_long flush(char *out,u_long sizeOUT);
+	u_long getHEADER(char *buffer,u_long buffersize);
 private:
 	gzip_data data;
 };
