+++ myserverweb/source/gzip.cpp
@@ -50,7 +50,7 @@ char GZIP_HEADER[]={(char)0x1f,(char)0x8b,Z_DEFLATED,0,0,0,0,0,0,0x03};
 /*!
 *Initialize the gzip structure value.
 */
-int Gzip::initialize()
+u_long Gzip::initialize()
 {
 #ifndef DO_NOT_USE_GZIP		
 	long level = Z_DEFAULT_COMPRESSION;
@@ -70,7 +70,7 @@ int Gzip::initialize()
 };
 
 #ifdef GZIP_CHECK_BOUNDS	
-int Gzip::compressBound(int size)
+u_long Gzip::compressBound(int size)
 {
 #ifdef compressBound 
 	return compressBound(size);
@@ -83,12 +83,12 @@ int Gzip::compressBound(int size)
 /*!
 *Compress the in buffer to the out buffer using the gzip compression.
 */
-int Gzip::compress(char* in,u_long sizeIn,char *out,u_long sizeOut)
+u_long Gzip::compress(char* in,u_long sizeIn,char *out,u_long sizeOut)
 {
 #ifndef DO_NOT_USE_GZIP	
 	u_long old_total_out=data.stream.total_out;
 	uLongf destLen=sizeOut;
-	int ret;
+	u_long ret;
 
 #ifdef GZIP_CHECK_BOUNDS
 	if(compressBound(sizeIn)>(u_long)sizeOut)
@@ -103,22 +103,22 @@ int Gzip::compress(char* in,u_long sizeIn,char *out,u_long sizeOut)
 
 	data.data_size+=data.stream.total_out-old_total_out;
 	data.crc = crc32(data.crc, (const Bytef *) in, sizeIn);
-	return static_cast<int>(data.stream.total_out-old_total_out);
+	return data.stream.total_out-old_total_out;
 #else 
 	/*
 	*If is specified DO_NOT_USE_GZIP copy the input buffer to the output one as it is.
 	*/
 	memcpy(out, in, min(sizeIn,sizeOut));
-	return min(static_cast<int>(sizeIn), static_cast<int>(sizeOut));
+	return min(sizeIn, sizeOut);
 #endif
 }
 
 /*!
  *Close the gzip compression.
  */
-int Gzip::free()
+u_long Gzip::free()
 {
-int ret=0;
+  u_long ret=0;
 #ifndef DO_NOT_USE_GZIP
 	
 	if(data.initialized==0)
@@ -131,7 +131,7 @@ int ret=0;
 /*!
 *Flush all the data
 */
-int Gzip::flush(char *out,u_long sizeOUT)
+u_long Gzip::flush(char *out,u_long sizeOUT)
 {
 #ifndef DO_NOT_USE_GZIP	
 	int ret;
@@ -146,7 +146,7 @@ int Gzip::flush(char *out,u_long sizeOUT)
 	ret = deflate(&(data.stream), Z_FINISH);
 
 	data.data_size+=data.stream.total_out-old_total_out;
-	return static_cast<int>(data.stream.total_out-old_total_out);
+	return data.stream.total_out-old_total_out;
 #else 
 	return 0;
 #endif
@@ -155,7 +155,7 @@ int Gzip::flush(char *out,u_long sizeOUT)
 /*!
  *Update the existent CRC.
  */
-int Gzip::updateCRC(char* buffer, int size)
+u_long Gzip::updateCRC(char* buffer, int size)
 {
 #ifndef DO_NOT_USE_GZIP		
 	data.crc = crc32(data.crc, (const Bytef *) buffer,(u_long)size);
@@ -168,7 +168,7 @@ int Gzip::updateCRC(char* buffer, int size)
 /*!
  *Get the GZIP footer.
  */
-int Gzip::getFOOTER(char *str,int /*size*/)
+u_long Gzip::getFOOTER(char *str,int /*size*/)
 {
 #ifndef DO_NOT_USE_GZIP		
 	char *footer =  str;
@@ -190,7 +190,7 @@ int Gzip::getFOOTER(char *str,int /*size*/)
 /*!
  *Copy the GZIP header in the buffer.
  */
-int Gzip::getHEADER(char *buffer,u_long buffersize)
+u_long Gzip::getHEADER(char *buffer,u_long buffersize)
 {
 	if(buffersize<GZIP_HEADER_LENGTH)
 		return 0;
